requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
comtypes==1.2.0
fastapi==0.92.0
Flask==2.3.2
frozenlist==1.4.0
gTTS==2.3.2
h11==0.14.0
httptools==0.6.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.8
PyAudio==0.2.13
pydantic==1.10.12
pypiwin32==223
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6
pyttsx3==2.90
pywin32==306
PyYAML==6.0.1
requests==2.28.2
sniffio==1.3.0
SpeechRecognition==3.10.0
starlette==0.25.0
tqdm==4.66.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.6
yarl==1.9.2